+++ src/uCtrlCore/Statistics/ustatistic.h
@@ -33,13 +33,13 @@ public:
     void data(const QString& data);
     int type() const;
     void type(int type);
-    uint timestamp() const;
-    void timestamp(uint timestamp);
+    double timestamp() const;
+    void timestamp(double timestamp);
 
 private:
     QString m_data;
     bool m_type;
-    uint m_timestamp;
+    double m_timestamp;
 };
 
 #endif // USTATISTIC_H
